syncflow_backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
python-dotenv==1.1.1
openai==1.102.0
pydantic==2.11.7
pydantic-settings==2.10.1
httpx==0.28.1
